skip some of our crontab hacks if we're using god

+++ config/schedule.rb
@@ -7,23 +7,26 @@
 
 set :cron_log, "#{RAILS_ROOT}/log/cron_log.log"
 set :server_type, nil
+set :have_god, nil
 
-case server_type
-when 'thin'
-  every 1.day, :at => "3:37 am" do
-    command "thin restart -C #{RAILS_ROOT}/config/thin.yml"
+unless have_god
+  case server_type
+  when 'thin'
+    every 1.day, :at => "3:37 am" do
+      command "thin restart -C #{RAILS_ROOT}/config/thin.yml"
+    end
+
+  when 'unicorn'
+    every 1.day, :at => "3:37 am" do
+      command "[ -f #{RAILS_ROOT}/tmp/pids/unicorn.pid ] && kill -HUP `cat #{RAILS_ROOT}/tmp/pids/unicorn.pid`"
+    end
   end
 
-when 'unicorn'
-  every 1.day, :at => "3:37 am" do
-    command "[ -f #{RAILS_ROOT}/tmp/pids/unicorn.pid ] && kill -HUP `cat #{RAILS_ROOT}/tmp/pids/unicorn.pid`"
+  every :reboot do
+    rake 'ts:start'
   end
 end
 
 every 8.hours, :at => 37 do
   rake "ts:in"
-end
-
-every :reboot do
-  rake 'ts:start'
 end
